US-8557 sql lock account_move_line before recreating AJI to prevent duplicates on concurrent actions

+++ bin/addons/account/account.py
@@ -1767,6 +1767,7 @@ class account_move(osv.osv):
         valid_moves = [] #Maintains a list of moves which can be responsible to create analytic entries
         obj_analytic_line = self.pool.get('account.analytic.line')
         obj_move_line = self.pool.get('account.move.line')
+        cr.execute('select id from account_move_line where move_id in %s for update', (tuple(ids),))
         for move in self.browse(cr, uid, ids, context):
             # Unlink old analytic lines on move_lines
             # UTP-803: condition on context added, if this context is set analytic lines won't be created,
@@ -1779,10 +1780,15 @@ class account_move(osv.osv):
                 valid_moves.append(move)
                 continue
             if not context.get('do_not_create_analytic_line') or not context.get('sync_update_execution'):
+                if move.state == 'posted':
+                    raise osv.except_osv(_('Error'), _('You cannot call this on a posted entry %s') % move.name)
                 for obj_line in move.line_id:
                     for obj in obj_line.analytic_lines:
                         obj_analytic_line.unlink(cr,uid,obj.id)
-
+            #import time
+            #print 'sleep...'
+            #time.sleep(10)
+            #print 'ok'
             journal = move.journal_id
             amount = 0
             amount_currency = 0
